+++ packages/shared_preferences/shared_preferences/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.3.5
+
+* Adds clarifying comment about allowList handling with an updated prefix.
+* Adds migration tool to move from legacy `SharedPreferences` to `SharedPreferencesAsync`.
+
 ## 2.3.3
 
 * Clarifies scope of prefix handling in README.